requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
pydantic==2.6.1
transformers==4.37.2
torch==2.2.0
python-multipart==0.0.9
Pillow==10.2.0
google-genai>=0.4.0
requests==2.31.0
beautifulsoup4==4.12.2
google-auth==2.29.0
httpx==0.27.0

